moved hashing structures to bench-templates crate, started ligero bench coding
