+++ poly-commit/Cargo.toml
@@ -26,8 +26,13 @@ rayon = { version = "1", optional = true }
 merlin = { version = "3.0.0", default-features = false }
 
 [[bench]]
-name = "pcs"
-path = "benches/pcs.rs"
+name = "ipa_times"
+path = "benches/ipa_times.rs"
+harness = false
+
+[[bench]]
+name = "ligero_ml_times"
+path = "benches/ligero_ml_times.rs"
 harness = false
 
 [[bench]]
