+++ poly-commit/src/linear_codes/mod.rs
@@ -31,8 +31,6 @@ mod ligero;
 use data_structures::*;
 
 pub use data_structures::{LigeroPCParams, LinCodePCProof};
-#[cfg(any(feature = "benches", test))]
-pub use utils::{FieldToBytesColHasher, LeafIdentityHasher};
 
 use utils::{calculate_t, get_indices_from_transcript, hash_column};
